+++ src/Factory/AuthorizeMiddlewareFactory.php
@@ -6,6 +6,7 @@ namespace Sirix\Mezzio\Rbac\Factory;
 
 use Psr\Container\ContainerExceptionInterface;
 use Psr\Container\ContainerInterface;
+use Sirix\ContainerResolver\ConfigReader;
 use Sirix\ContainerResolver\ContainerResolver;
 use Sirix\Mezzio\Rbac\Contract\RequestGuardInterface;
 use Sirix\Mezzio\Rbac\Middleware\AuthorizeMiddleware;
@@ -18,9 +19,11 @@ final class AuthorizeMiddlewareFactory
     public function __invoke(ContainerInterface $container): AuthorizeMiddleware
     {
         $containerResolver = ContainerResolver::forFactory($container, self::class);
+        $configReader      = ConfigReader::fromContainer($containerResolver);
 
         return new AuthorizeMiddleware(
             $containerResolver->get(RequestGuardInterface::class),
+            $configReader->bool('rbac.authorize_middleware.strict', true),
         );
     }
 }